add test for mapping
